requirements/core.txt
Flask==1.1.2
flask-restx==0.2.0
Flask-SQLAlchemy==2.4.1
